refactor(view): MVC 패턴 공부, 적용중
