+++ src/main/java/lotto/Utils/Validator/Validation.java
@@ -1,5 +1,7 @@
-package lotto;
+package lotto.Utils.Validator;
 
+
+import lotto.View.OutputView;
 
 import java.util.ArrayList;
 import java.util.HashSet;
@@ -10,7 +12,7 @@ public class Validation {
 
     public static void Integer(String string) {
         if (!string.matches("-?\\d+")) {
-            View.errorMessage("입력값이 옳지 않습니다.");
+            OutputView.errorMessage("입력값이 옳지 않습니다.");
             throw new IllegalArgumentException();
         }
     }
@@ -18,7 +20,7 @@ public class Validation {
     public static void validateRange(String StringNumber) {
         int number = Integer.parseInt(StringNumber);
         if (!(number > 0 && number < 46)) {
-            View.errorMessage("입력값의 범위가 맞지 않습니다..");
+            OutputView.errorMessage("입력값의 범위가 맞지 않습니다..");
             throw new IllegalArgumentException();
         }
     }
@@ -28,7 +30,7 @@ public class Validation {
         Set<Integer> numSet = new HashSet<>(List);
 
         if (numSet.size() != List.size()) {
-            View.errorMessage("중복되는 값이 있습니다.");
+            OutputView.errorMessage("중복되는 값이 있습니다.");
             throw new IllegalArgumentException();
         }
     }
@@ -44,14 +46,14 @@ public class Validation {
 
     public static void length(List<Integer> numbers) {
         if (numbers.size() != 6) {
-            View.errorMessage("당첨번호의 길이가 올바르지 않습니다.");
+            OutputView.errorMessage("당첨번호의 길이가 올바르지 않습니다.");
             throw new IllegalArgumentException();
         }
     }
 
     public static void price(int price) {
         if (price % 1000 != 0) {
-            View.errorMessage("금액이 올바르지 않습니다.");
+            OutputView.errorMessage("금액이 올바르지 않습니다.");
             throw new IllegalArgumentException();
         }
     }
